FakeNewsExtension/server/requirements.txt
newspaper3k
lxml_html_clean
selenium
webdriver_manager
fastapi
uvicorn 
pydantic 
langchain 
langchain-google-genai